% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_FvCB1980.R
\name{FvCB1980}
\alias{FvCB1980}
\title{Calculate carboxylation rate limited by three process
Wc: RuBP-saturated carboxylation rate
Wj: RuBP regeneration-limited carboxylation rate
Wp: TPU-limited carboxylation rate}
\usage{
FvCB1980(envs, pars, Ct = NULL)
}
\arguments{
\item{envs}{List of environmental variables:
\describe{
\item{\code{PPFD}}{Photosynthetic photon flux density in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{C_chl}}{CO2 concentration in chloroplast in μmol mol\eqn{^{-1}}).}
}}

\item{pars}{List of parameters:
\describe{
\item{\code{V_cmax}}{Maximum rate of RuBP-saturated carboxylation in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{J_max}}{Maximum rate of electron transport through PSII in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{V_tpu}}{Rate of triose phosphate export from the chloroplast in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{K_CO}}{Apparent Michaelis-Menten constant of Rubisco, same units as CO2.}
\item{\code{R_d}}{Day respiration in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{gamma_star}}{CO2 photocompensation point, same units as CO2.}
\item{\code{phi_J}}{quantum efficiency, unitless.}
\item{\code{theta_J}}{Convexity factor for response of J to light, unitless.}
}}

\item{Ct}{Optional input (default = \code{NULL}). A numeric vector of two element:
first is Ci transition between the Rubisco and RuBP Regeneration limitations
second is Ci transition between the RuBP Regeneration and TPU limitations}
}
\value{
List of calculated rate:
\describe{
\item{\code{W_c}}{RuBP-saturated carboxylation rate in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{W_j}}{RuBP regeneration-limited carboxylation rate in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{W_p}}{TPU-limited carboxylation rate in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{Wmin}}{Minimum between W_c, W_j and W_p.}
\item{\code{An}}{Net assimilation in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{Ac}}{Net assimilation, RuBP-saturated, in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{Aj}}{Net assimilation, RuBP regeneration-limited, in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{Ap}}{Net assimilation, TPU-limited, in μmol m\eqn{^{-2}} s\eqn{^{-1}}.}
\item{\code{Limitation}}{Character vector specifying the limitation at which CO2 level.}
}
}
\description{
Calculate carboxylation rate limited by three process
Wc: RuBP-saturated carboxylation rate
Wj: RuBP regeneration-limited carboxylation rate
Wp: TPU-limited carboxylation rate
}
